Avoid recreating config with no uuid

See merge request project/drupal!8354

+++ core/lib/Drupal/Core/Config/ConfigImporter.php
@@ -964,8 +964,11 @@ class ConfigImporter {
             $entity_storage = $this->configManager->getEntityTypeManager()->getStorage($entity_type_id);
             $entity_type = $this->configManager->getEntityTypeManager()->getDefinition($entity_type_id);
             $entity = $entity_storage->load($entity_storage->getIDFromConfigName($name, $entity_type->getConfigPrefix()));
-            $entity->delete();
-            $this->logError($this->t('Deleted and replaced configuration entity "@name"', ['@name' => $name]));
+
+            if ($entity) {
+              $entity->delete();
+              $this->logError($this->t('Deleted and replaced configuration entity "@name"', ['@name' => $name]));
+            }
           }
           else {
             $this->storageComparer->getTargetStorage($collection)->delete($name);
