+++ core/lib/Drupal/Core/Config/Entity/ConfigEntityStorage.php
@@ -362,8 +362,12 @@ class ConfigEntityStorage extends EntityStorageBase implements ConfigEntityStora
   public function importDelete($name, Config $new_config, Config $old_config) {
     $id = static::getIDFromConfigName($name, $this->entityType->getConfigPrefix());
     $entity = $this->load($id);
-    $entity->setSyncing(TRUE);
-    $entity->delete();
+
+    if ($entity) {
+      $entity->setSyncing(TRUE);
+      $entity->delete();
+    }
+
     return TRUE;
   }
